third_party/WebKit/LayoutTests/http/tests/security/postMessage/invalid-origin-throws-exception-expected.txt

CONSOLE MESSAGE: Unable to post message to asdf://. Recipient has origin http://localhost:8000.

window.location.href = http://127.0.0.1:8000/security/postMessage/invalid-origin-throws-exception.html
waiting...
Encountered exception SyntaxError: An invalid or illegal string was specified. while posting message to ''.
Encountered exception SyntaxError: An invalid or illegal string was specified. while posting message to 'asdf'.
Encountered exception SyntaxError: An invalid or illegal string was specified. while posting message to '/tmp/foo'.
Encountered exception SyntaxError: An invalid or illegal string was specified. while posting message to '//localhost'.
Posted message to 'asdf:' without any exceptions.
Encountered exception SyntaxError: An invalid or illegal string was specified. while posting message to 'http:'.
Received message: data="Received message: data="done" origin="http://127.0.0.1:8000"" origin="http://localhost:8000"
